feat(web): implement navbar in all pages
