+++ apps/web/src/modules/dashboard/components/dashboard-content.tsx
@@ -1,12 +1,7 @@
 "use client";
 
-import { PlusIcon } from "lucide-react";
-import { useTranslations } from "next-intl";
-import { Button } from "@/components/ui/button";
-import { Separator } from "@/components/ui/separator";
-import { SheetTrigger } from "@/components/ui/sheet";
 import { cn } from "@/lib/utils";
-import { CreateReport } from "@/modules/report";
+import { DashboardNavbar } from "./dashboard-navbar";
 import { DashboardReportList } from "./dashboard-report-list";
 import { DashboardStats } from "./dashboard-stats";
 
@@ -16,9 +11,9 @@ function DashboardContent({
 }: React.ComponentProps<"div">) {
 	return (
 		<div className={cn("", className)} data-slot="dashboard-content" {...props}>
+			<DashboardNavbar />
 			<main className="py-12">
-				<DashboarHeader />
-				<section className="container mt-8">
+				<section className="container">
 					<DashboardStats />
 				</section>
 				<section className="container mt-20">
@@ -26,35 +21,6 @@ function DashboardContent({
 				</section>
 			</main>
 		</div>
-	);
-}
-
-function DashboarHeader({
-	className,
-	...props
-}: React.ComponentProps<"section">) {
-	const t = useTranslations("modules.dashboard.header");
-
-	return (
-		<section
-			className={cn("container", className)}
-			data-slot="dashboard-header"
-			{...props}
-		>
-			<div className="flex flex-wrap justify-between gap-4">
-				<h1 className="font-semibold text-2xl text-slate-800">{t("title")}</h1>
-				<CreateReport>
-					<SheetTrigger
-						render={
-							<Button size={"sm"}>
-								<PlusIcon /> {t("newReport")}
-							</Button>
-						}
-					/>
-				</CreateReport>
-			</div>
-			<Separator className={"mt-4"} />
-		</section>
 	);
 }
