+++ apps/web/src/modules/report/components/report-content.tsx
@@ -4,6 +4,7 @@ import { ReportAttachments } from "./report-attachments";
 import { ReportDetails } from "./report-details";
 import { ReportExpenses } from "./report-expenses";
 import { ReportHeader } from "./report-header";
+import { ReportNavbar } from "./report-navbar";
 import { ReportPaidNotice } from "./report-paid-notice";
 
 function ReportContent({
@@ -14,7 +15,8 @@ function ReportContent({
 	reportId: string;
 }) {
 	return (
-		<div className={cn("py-12", className)} data-slot="report-content" {...props}>
+		<div className={cn("pb-12", className)} data-slot="report-content" {...props}>
+			<ReportNavbar className="mb-12" reportId={reportId} />
 			<div className="container mb-8">
 				<ReportPaidNotice reportId={reportId} />
 			</div>
