+++ apps/web/src/modules/reports/components/reports-content.tsx
@@ -1,9 +1,11 @@
 import { cn } from "@/lib/utils";
 import { ReportsList } from "./list";
+import { ReportsNavbar } from "./reports-navbar";
 
 function ReportsContent({ className, ...props }: React.ComponentProps<"div">) {
 	return (
 		<div className={cn("", className)} data-slot="component" {...props}>
+			<ReportsNavbar className="mb-4" />
 			<ReportsList />
 		</div>
 	);
