+++ apps/web/src/modules/reports/components/reports-navbar.tsx
@@ -1,0 +1,24 @@
+import {
+	Breadcrumb,
+	BreadcrumbItem,
+	BreadcrumbList,
+	BreadcrumbPage,
+} from "@zemio/ui";
+import { Navbar } from "@/components/navbar";
+import { cn } from "@/lib/utils";
+
+function ReportsNavbar({ className, ...props }: React.ComponentProps<"div">) {
+	return (
+		<Navbar className={cn("", className)} data-slot="reports-navbar" {...props}>
+			<Breadcrumb>
+				<BreadcrumbList>
+					<BreadcrumbItem>
+						<BreadcrumbPage>Reports</BreadcrumbPage>
+					</BreadcrumbItem>
+				</BreadcrumbList>
+			</Breadcrumb>
+		</Navbar>
+	);
+}
+
+export { ReportsNavbar };
