+++ apps/web/src/modules/shared/components/app-navbar.tsx
@@ -1,10 +1,10 @@
 "use client";
 
+import { SidebarTrigger } from "@zemio/ui";
 import { CirclePlusIcon, LifeBuoyIcon, SettingsIcon } from "lucide-react";
 import Link from "next/link";
 import { Button, buttonVariants } from "@/components/ui/button";
 import { SheetTrigger } from "@/components/ui/sheet";
-import { SidebarTrigger } from "@/components/ui/sidebar";
 import { ROUTES } from "@/lib/routes";
 import { cn } from "@/lib/utils";
 import { CreateReport } from "@/modules/report";
